Revert "demo: improve demos (#45)"
This reverts commit b99bc69ab4083a299fd36a9fac0ae6964c533ffa.
