+++ demo/web/tsconfig.json
@@ -17,9 +17,9 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "paths": {
-      "@utils/*": [
-        "../utils/*"
-      ],
+      "@lib/*": [
+        "../../../lib/*"
+      ]
     }
   },
   "include": [
