+++ lib/index.ts
@@ -2,6 +2,7 @@
  * Index file
  */
 
+export * from "./focusable";
 export * from "./keyCombo";
 export * from "./keyMap";
 export * from "./navigation";
